[refactor]: 상품 가격, 수량 타입 변경
[목적]: 타 도메인과 타입을 맞추기 위함

[목표]: 상품 타입을 Long으로 변경

[달성도]:
  - 일괄 타입 변경 완료

[기타]: int를 Long으로 수정

+++ src/main/java/com/teamsparta14/order_service/product/dto/ProductResponseDto.java
@@ -15,8 +15,8 @@ public class ProductResponseDto {
     private UUID storeId;
     private UUID productId;
     private String productName;
-    private int productPrice;
-    private int productQuantity;
+    private Long productPrice;
+    private Long productQuantity;
 
     public static ProductResponseDto of(Product product) {
         return ProductResponseDto.builder()
